requirements.txt
arrow==0.15.2
blessed==1.15.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-admin==2.0.0
django-allauth==0.40.0
django-cors-headers==3.1.0
django-excel-response2==2.0.8
django-picklefield==2.0
django-redis==4.10.0
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
screen==1.0.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
xlwt==1.3.0
